Add files via upload
- data_preparation.py: reads the data from the matches and calculates the features as ranking position, average values of the received yellow cards, goals and others for the two teams. The features, together with the scores of the two playing teams will be saved in a DataFrame and saved locally with help of pickle.
This function is called by Prediction_model.py only in case the preparation has not already been done and the data not been saved

- Prediction_model.py: is divided in two parts
1. A random forest regressor is trained on the prepared data and with its help the most important features are found
2. The information about the most important features from the previous model are used to reduce the feature set. A MLP regressor is then trained on this reduced dataset. With help of a grid search the best parameters for this algorithm are found

In both cases the models are evaluated by calculating the winning team from the predicted scores, and comparing this result to the actual result. MLP regressor with the best parameters and most important features reaches a prediction accuracy of about 55%, 5% more than the RF regressor with best parameters and all features.